+++ CHANGELOG.md
@@ -1172,6 +1172,7 @@ TV-9.4.0.20220719
     - #066324 - Only a success message appears but the product does not get removed from the cart when the user clicks on the remove item from the cart button
     - #066380 - A duplicate entry error occurs when the user clicks on the reset button multiple times on the billing address setup popup.
     - #066357 - The text appears 'Cash On Delivery Available' on the all sellers page even if the COD payment method is disabled
+    - #066347 - sorting issue in category requests and brand requests
 
 ## UAT points: